EMG-RNN/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.9.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
greenlet==2.0.2
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lightning-utilities==0.8.0
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.2.1
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.1
optuna==3.1.1
packaging==23.1
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.2.0
plotly==5.14.1
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==2.0.1.post0
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.9
stack-data==0.6.2
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchmetrics==0.11.4
torchvision==0.15.1+cu118
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.6
yarl==1.8.2
